+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### = 3.2.7 (March 5, 2019) =
 
+* New: Added a new form type which is dynamical fields.
 * Tweak: Improve the sort object then sub-sort further for Dynamic Drop Down List from JSON.
 * Tweak: Improve the form components styles.
 * Tweak: Upgraded the Date Picker function.